requirements.txt
flask
authlib
SQLAlchemy
python-dotenv